cambios deplou

+++ src/infrastructure/repositories/ProductoRepository.js
@@ -54,9 +54,9 @@ class ProductoRepository extends IProductoRepository {
       estado: r.estado,
       ciudad_id: parseInt(r.ciudad_id) || 0,
       nombre_ciudad: r.nombre_ciudad,
-      total_compras: parseInt(r.total_compras) || 0,
-      total_ventas: parseInt(r.total_ventas) || 0,
-      saldo: parseInt(r.saldo) || 0,
+      total_compras: parseFloat(r.total_compras) || 0,
+      total_ventas: parseFloat(r.total_ventas) || 0,
+      saldo: parseFloat(r.saldo) || 0,
       stock_bajo: r.stock_bajo === true || r.stock_bajo === 'true',
     }));
   }
